Iterator test Migration

+++ proj.win32/Game.cpp
@@ -7,17 +7,58 @@
 #include <proj.win32/Level2Scene.h>
 #include <proj.win32/Level3Scene.h>
 #include <proj.win32/GeneratorCard.h>
+#include <proj.win32/CardIterator.h>
 #include <Math.h>
 
 USING_NS_CC;
 
 void Game::Turn(position pos, Card* cards[3][3], int level) {
-	if (abs(pos.x - gameData::heroPosition.x) + abs(pos.y - gameData::heroPosition.y) == 1) {
+	if (abs(pos.x - gameData::heroPosition.x) + abs(pos.y - gameData::heroPosition.y) == 1 && !gameData::isSceneLocked) {
 		if (cards[pos.x][pos.y]->cardInteract() == 1) {
+			std::vector<position> cardVector;
+			CardIterator cardIterator(gameData::heroPosition, pos);
+
+			//Lock Scene
+			gameData::isSceneLocked = true;
+
+
 			cards[pos.x][pos.y]->deleteCard();
-			cards[pos.x][pos.y] = cards[gameData::heroPosition.x][gameData::heroPosition.y];
-			cards[pos.x][pos.y]->moveCard(position(pos.x - gameData::heroPosition.x, pos.y - gameData::heroPosition.y));
-			cards[gameData::heroPosition.x][gameData::heroPosition.y] = GeneratorCard(1, gameData::currentScene).GenerateRandomCard(*(new position(gameData::heroPosition.x, gameData::heroPosition.y)));
+
+
+			cardVector.push_back(pos);
+			while (cardIterator.hasNext())
+			{
+				cardVector.push_back(cardIterator.getNext());
+			}
+
+			//Card*[][] Replace
+			for (int i = 0; i < cardVector.size() - 1; i++)
+			{
+				cards[cardVector[i].x][cardVector[i].y] = cards[cardVector[i + 1].x][cardVector[i + 1].y];
+			}
+
+			//Creating CallBack Functions
+			position endPos = cardVector[cardVector.size() - 1];
+			auto createCB = CallFunc::create([endPos, cardVector, cards, level]() {
+				cards[endPos.x][endPos.y] = GeneratorCard(level, gameData::currentScene).GenerateRandomCard(*(new position(endPos)));
+				});
+
+			auto unlockCB = CallFunc::create([endPos, cardVector, cards, level]() {
+				gameData::isSceneLocked = false;
+				});
+
+			//Card Move With Delays
+			for (int i = 0; i < cardVector.size() - 1; i++)
+			{
+				position vecDirection = cardVector[i] - cardVector[i + 1];
+				cards[cardVector[i].x][cardVector[i].y]->spriteCard->runAction(Sequence::create(DelayTime::create(i * 0.75), MoveBy::create(0.75, Vec2(vecDirection.y * (192 + 10), -vecDirection.x * (192 + 10)))->clone(), nullptr));
+				cards[cardVector[i].x][cardVector[i].y]->spriteFrame->runAction(Sequence::create(DelayTime::create(i * 0.75), MoveBy::create(0.75, Vec2(vecDirection.y * (192 + 10), -vecDirection.x * (192 + 10)))->clone(), nullptr));
+				if (i == cardVector.size() - 2)
+				{
+					cards[cardVector[i].x][cardVector[i].y]->spriteFrame->runAction(Sequence::create(DelayTime::create((i + 1) * 0.75 + 0.15), createCB, unlockCB, nullptr));
+				}
+			}
+
 			gameData::heroPosition = position(pos);
 		}
 	}
